+++ client/Packages/com.beamable/Tests/Runtime/Beamable/BeamableTest.cs
@@ -12,6 +12,9 @@ namespace Beamable.Tests.Runtime
 	/// </summary>
 	public class BeamableTest
 	{
+		public const string DEFAULT_PID = "testPid";
+		public const string DEFAULT_CID = "123";
+		
 		protected MockPlatformAPI MockRequester;
 		protected User MockPlatformUser;
 		protected MockBeamableApi MockApi;
@@ -27,7 +30,7 @@ namespace Beamable.Tests.Runtime
 			MockPlatformUser = new User { id = 12 };
 			MockPlatform.User = MockPlatformUser;
 			MockApi.User = MockPlatform.User;
-			MockApi.Token = new AccessToken(null, "123", "testPid", "testtoken", "refresh", 0);
+			MockApi.Token = new AccessToken(null, DEFAULT_CID, DEFAULT_PID, "testtoken", "refresh", 0);
 			MockRequester = new MockPlatformAPI();
 			MockRequester.Token = MockApi.Token;
 			MockApi.Requester = MockRequester;