+++ client/Packages/com.beamable/CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Broken Console commands, `config set`, `config reset`, and `config useful`.
+- `ConfigDatabase` is deprecated.  
 
 ## [1.18.0]